requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
ShopifyAPI==9.0.0
certifi==2023.7.22
pytest==7.4.0